+++ src/main/java/com/example/docconneting/common/config/RabbitMQConfig.java
@@ -27,6 +27,33 @@ public class RabbitMQConfig {
     @Value("${alarm.queue.name}")
     private String alarmQueueName;
 
+    @Value("${rabbitmq.queue.coupon-name}")
+    private String couponQueueName;
+
+    @Value("${rabbitmq.queue.dlq-name}")
+    private String dlqQueueName;
+
+    @Value("${rabbitmq.queue.retry-name}")
+    private String retryQueueName;
+
+    @Value("${rabbitmq.queue.fail-name}")
+    private String failQueueName;
+
+    @Value("${rabbitmq.exchange.name}")
+    private String exchangeName;
+
+    @Value("${rabbitmq.routing.key}")
+    private String routingKey;
+
+    @Value("${rabbitmq.routing.dlq-key}")
+    private String dlqRoutingKey;
+
+    @Value("${rabbitmq.routing.retry-key}")
+    private String retryRoutingKey;
+
+    @Value("${rabbitmq.routing.fail-key}")
+    private String failRoutingKey;
+
     @Bean
     public Queue queue(){
         return new Queue(queue + "." + id);
@@ -43,8 +70,39 @@ public class RabbitMQConfig {
     }
 
     @Bean
+    public Queue couponQueue() {
+        return QueueBuilder.durable(couponQueueName)
+                .withArgument("x-dead-letter-exchange", exchangeName)
+                .withArgument("x-dead-letter-routing-key", dlqRoutingKey)
+                .build();
+    }
+
+    // DLQ (죽은 메세지 큐)
+    @Bean
+    public Queue dlqQueue() {
+        return QueueBuilder.durable(dlqQueueName).build();
+    }
+
+    // 재시도 큐
+    @Bean
+    public Queue retryQueue() {
+        return QueueBuilder.durable(retryQueueName).build();
+    }
+
+    // 최종 실패 정보 저장 큐
+    @Bean
+    public Queue failQueue() {
+        return QueueBuilder.durable(failQueueName).build();
+    }
+
+    @Bean
     public FanoutExchange exchange(){
         return new FanoutExchange(exchange);
+    }
+
+    @Bean
+    public DirectExchange couponExchange() {
+        return new DirectExchange(exchangeName);
     }
 
     @Bean
@@ -60,6 +118,30 @@ public class RabbitMQConfig {
                 .bind(elasticsearchQueue())
                 .to(exchange());
     }
+
+    @Bean
+    public Binding couponBinding(Queue queue, DirectExchange exchange) {
+        return BindingBuilder.bind(queue).to(exchange).with(routingKey);
+    }
+
+    // 쿠폰 DLQ 바인딩
+    @Bean
+    public Binding dlqBinding() {
+        return BindingBuilder.bind(dlqQueue()).to(couponExchange()).with(dlqRoutingKey);
+    }
+
+    // 쿠폰 재시도 큐 바인딩
+    @Bean
+    public Binding retryBinding() {
+        return BindingBuilder.bind(retryQueue()).to(couponExchange()).with(retryRoutingKey);
+    }
+
+    // 쿠폰 실패 큐 바인딩
+    @Bean
+    public Binding failBinding() {
+        return BindingBuilder.bind(failQueue()).to(couponExchange()).with(failRoutingKey);
+    }
+
 
     @Bean
     public MessageConverter converter(){
